Save error messages

+++ lib/notifiable/apns/apnotic/async.rb
@@ -27,11 +27,9 @@ module Notifiable
           push.on(:response) do |response|
             if response.ok?
               processed(device, 0)
-            elsif response.status == '410' || (response.status == '400' && response.body['reason'] == 'BadDeviceToken')
-              device.destroy
-              processed(device, response.status)
             else
-              processed(device, response.status)
+              processed(device, response.status, response.body['reason'])
+              device.destroy if response.status == '410' || (response.status == '400' && response.body['reason'] == 'BadDeviceToken')
             end
           end
